+++ testdata/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest/com/liferay/portal/kernel/service/MyCustomService.java
@@ -1,0 +1,5 @@
+package com.liferay.portal.kernel.service;
+
+public class MyCustomService implements com.liferay.portal.kernel.service.BaseLocalService {
+
+}